updtated

+++ Backend/AirlineReservationSystem/src/main/java/com/siva/AirlineReservationSystem/repository/BookingRepository.java
@@ -3,5 +3,4 @@ package com.siva.AirlineReservationSystem.repository;
 import com.siva.AirlineReservationSystem.entity.Booking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookingRepository extends JpaRepository<Booking, Integer> {
-}
+public interface BookingRepository extends JpaRepository<Booking, Long> {}
